Refactor nodes presentation
